requirement.txt
streamlit
langchain 
langchain-core 
langchain-community 
langchain-google-genai 
python-dotenv 
chromadb 
sentence-transformers 
beautifulsoup4 
requests 
faiss-cpu